+++ src/components/AddGuest.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const AddGuest = () => {
+    return (
+        <React.Fragment>
+            
+        </React.Fragment>
+    )
+}
+
+export default AddGuest;
